requirements.txt
# Dependencies for Real Estate Sentinel App

# Core Streamlit framework
streamlit>=1.28.0

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.23.0

# Visualization
plotly>=5.14.0

# Configuration file handling
pyyaml>=6.0

# Financial data APIs
fredapi>=0.5.0
yfinance>=0.2.0
